requirements.txt
asgiref==3.8.1
Django==5.1.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-nested-routers==0.94.1
drf-yasg==1.21.8
inflection==0.5.1
packaging==24.2
PyJWT==2.10.1
pytz==2024.2
PyYAML==6.0.2
shortuuid==1.0.13
sqlparse==0.5.3
tzdata==2024.2
uritemplate==4.1.1
